Imported Upstream version 5.0~beta1a

+++ src/RageUtil.cpp
@@ -43,7 +43,7 @@ void MersenneTwister::Reset( int iSeed )
 
 void MersenneTwister::GenerateValues()
 {
-	static const int mask[] = { 0, 0x9908B0DF };
+	static const unsigned mask[] = { 0, 0x9908B0DF };
 
 	for( int i = 0; i < 227; ++i )
 	{
@@ -98,8 +98,8 @@ namespace
 {
 	MersenneTwister g_LuaPRNG;
 
-	/* To map from [0..2^32-1] to [0..1), we divide by 2^32. */
-	const double DIVISOR = pow( double(2), double(32) );
+	/* To map from [0..2^31-1] to [0..1), we divide by 2^31. */
+	const double DIVISOR = pow( double(2), double(31) );
 
 	static int Seed( lua_State *L )
 	{
